# server/swagger.yaml
openapi: 3.0.0
info:
  title: Kanata Scraper API
  description: API documentation for Kanata Scraper (Otakudesu, etc.)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/otakudesu/complete:
    get:
      tags:
        - Otakudesu
      summary: Get Otakudesu Complete Anime List
      description: Scrapes the list of completed anime from Otakudesu with pagination.
      parameters:
        - in: query
          name: page
          schema:
            type: string
          description: Page number (default is 1)
      responses:
        '200':
          description: Complete anime list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title: { type: string }
                    episode: { type: string }
                    score: { type: string }
                    date: { type: string }
                    thumb: { type: string }
                    slug: { type: string }
  /api/otakudesu/ongoing:
    get:
      tags:
        - Otakudesu
      summary: Get Otakudesu Ongoing Anime List
      description: Scrapes the list of ongoing anime from Otakudesu with pagination.
      parameters:
        - in: query
          name: page
          schema:
            type: string
          description: Page number (default is 1)
      responses:
        '200':
          description: Ongoing anime list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title: { type: string }
                    episode: { type: string }
                    day: { type: string }
                    date: { type: string }
                    thumb: { type: string }
                    slug: { type: string }
  /api/otakudesu/search:
    get:
      tags:
        - Otakudesu
      summary: Search Anime
      description: Searches for anime on Otakudesu using a query string.
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: true
          description: Search query (e.g., one piece)
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title: { type: string }
                    thumb: { type: string }
                    genres: { type: array, items: { type: string } }
                    status: { type: string }
                    rating: { type: string }
                    slug: { type: string }
  /api/otakudesu/anime/{slug}:
    get:
      tags:
        - Otakudesu
      summary: Get Anime Details
      description: Scrapes details of a specific anime and its episode list.
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
          description: Anime slug (e.g., 1piece-sub-indo)
      responses:
        '200':
          description: Anime details and episode list
          content:
            application/json:
              schema:
                type: object
                properties:
                  title: { type: string }
                  japanese: { type: string }
                  score: { type: string }
                  producer: { type: string }
                  type: { type: string }
                  status: { type: string }
                  total_episode: { type: string }
                  duration: { type: string }
                  release_date: { type: string }
                  studio: { type: string }
                  genres: { type: array, items: { type: string } }
                  thumb: { type: string }
                  synopsis: { type: string }
                  batch_link:
                    type: object
                    nullable: true
                    properties:
                      title: { type: string }
                      slug: { type: string }
                  episode_list:
                    type: array
                    items:
                      type: object
                      properties:
                        title: { type: string }
                        slug: { type: string }
                        date: { type: string }
  /api/otakudesu/batch/{slug}:
    get:
      tags:
        - Otakudesu
      summary: Get Batch Download Links
      description: Scrapes batch download links for a specific anime batch.
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
          description: Batch slug (e.g., wpoiec-batch-sub-indo)
      responses:
        '200':
          description: Batch details and download links
          content:
            application/json:
              schema:
                type: object
                properties:
                  title: { type: string }
                  thumb: { type: string }
                  download_list:
                    type: array
                    items:
                      type: object
                      properties:
                        title: { type: string }
                        links:
                          type: array
                          items:
                            type: object
                            properties:
                              quality: { type: string }
                              size: { type: string }
                              links:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    name: { type: string }
                                    url: { type: string }
...
  /api/otakudesu/genres/{slug}:
    get:
      tags:
        - Otakudesu
      summary: Get Anime by Genre
      description: Scrapes the list of anime for a specific genre with pagination.
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
          description: Genre slug (e.g., action)
        - in: query
          name: page
          schema:
            type: string
          description: Page number (default is 1)
      responses:
        '200':
          description: Anime list by genre
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title: { type: string }
                    studio: { type: string }
                    eps: { type: string }
                    rating: { type: string }
                    thumb: { type: string }
                    slug: { type: string }
                    date: { type: string }
                    genres: { type: array, items: { type: string } }
...
  /api/otakudesu/episode/{slug}:
    get:
      tags:
        - Otakudesu
      summary: Get Episode Details
      description: Scrapes an episode page to get the title, default stream, and mirror list.
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
          description: The episode slug (e.g., tamon-kun-episode-1-sub-indo)
      responses:
        "200":
          description: Episode details
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  defaultStream:
                    type: string
                    description: URL of the default iframe stream
                  mirrors:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: vidhide
                        payload:
                          type: object
                          description: JSON payload needed for /stream endpoint
        "500":
          description: Scraping failed
  /api/otakudesu/stream:
    post:
      tags:
        - Otakudesu
      summary: Resolve Stream URL
      description: Resolves the actual iframe URL for a specific mirror using the payload from the episode details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payload
              properties:
                payload:
                  type: object
                  description: The payload object obtained from the /episode/{slug} endpoint
                  example:
                    id: 191485
                    i: 0
                    q: "360p"
      responses:
        "200":
          description: Stream URL resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  iframe:
                    type: string
                    description: The resolved iframe URL
                  html:
                    type: string
                    description: The raw HTML containing the iframe (decoded)
        "400":
          description: Missing payload
        "500":
          description: Failed to fetch stream